requirements.txt
# Eärendil - Kerr Black Hole Visualizer
# =====================================

# Core
jax[cuda12]>=0.4.20
diffrax>=0.4.0
numpy>=1.24.0

# GUI
PyQt6>=6.4.0

# Sky data
requests>=2.28.0
Pillow>=9.0.0

# For CPU-only systems, replace jax[cuda12] with:
# jax>=0.4.20
